+++ apps/synth_desktop/src/renderer/src/components/mander/index.ts
@@ -1,0 +1,6 @@
+export { Mander } from "./Mander";
+export { ManderLab, ManderLabGate } from "./ManderLab";
+export { ManderPresence } from "./ManderPresence";
+export { resolveManderEmotion, sessionHasOpenTools } from "./Mander.presence";
+export type { ManderMotion, ManderProps, ManderState } from "./Mander.types";
+export { MANDER_STATES } from "./Mander.types";